special handling for fixwrapper
