+++ org.mwc.debrief.lite/src/main/java/org/mwc/debrief/lite/outline/OutlinePanelView.java
@@ -54,6 +54,7 @@ import javax.swing.JButton;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JTree;
+import javax.swing.SwingUtilities;
 import javax.swing.UIManager;
 import javax.swing.border.Border;
 import javax.swing.event.TreeSelectionEvent;
@@ -577,6 +578,82 @@ public class OutlinePanelView extends SwingLayerManager implements
     
 
   }
+  
+  /**
+   * have a fresh pass through the data
+   */
+  public void updateData(final Layer changedLayer, final Plottable newItem)
+  {
+    _myTree.setExpandsSelectedPaths(true);
+    // find out which node is currently visible
+    if (changedLayer != null && newItem != null)
+    {
+      DefaultMutableTreeNode rootNode = getTreeNode(null, changedLayer
+          .getName(), changedLayer);
+      if(rootNode!=null) {
+        final DefaultMutableTreeNode itemNode;
+        if(rootNode.getUserObject() instanceof Layer){
+          //used class name instead of instanceof as otherwise
+          //there will be cyclic dependency in eclipse manifest.
+         
+          if(newItem instanceof FixWrapper) {
+            rootNode = (DefaultMutableTreeNode)rootNode.getFirstChild();
+          }
+          itemNode = getTreeNode(rootNode, newItem.getName(),
+        
+            newItem);
+        }
+        else {
+          //if not a layer but a trackwrapper
+          DefaultMutableTreeNode firstChild = (DefaultMutableTreeNode)rootNode.getFirstChild();
+          if(firstChild != null) {
+            itemNode = getTreeNode((DefaultMutableTreeNode)rootNode.getFirstChild(), newItem.getName(),
+              
+              newItem);
+          }
+          else {
+            itemNode = null;
+          }
+        }
+        if(itemNode!=null) {
+          final TreePath _treePath = new TreePath(itemNode.getPath());
+          SwingUtilities.invokeLater(new Runnable()
+          {
+
+            @Override
+            public void run()
+            {
+              _myTree.expandPath(_treePath);
+              _myTree.scrollPathToVisible(_treePath);
+              _myTree.makeVisible(_treePath);
+              _myTree.setSelectionPath(_treePath);
+            }
+          });
+        }
+      }
+    }
+    else if (changedLayer != null)
+    {
+      final DefaultMutableTreeNode rootNode = getTreeNode(null, changedLayer
+          .getName(), changedLayer);
+      if(rootNode!=null) {
+        final TreePath _treePath = new TreePath(rootNode.getPath());
+        SwingUtilities.invokeLater(new Runnable()
+        {
+
+          @Override
+          public void run()
+          {
+            _myTree.expandPath(_treePath);
+            _myTree.scrollPathToVisible(_treePath);
+            _myTree.makeVisible(_treePath);
+            _myTree.setSelectionPath(_treePath);
+          }
+        });
+      }
+    }
+  }
+
 
   private void addBackData(Plottable theData, CanEnumerate destination)
   {
